Use nested loop over stripes
Although performance doesn't change (aside from the exception below), I
think this way of expressing the loop is more amenable to experimentation.

(For example, we can try adding loop pragmas to the two inner loops --
although this did not seem to have any benefit on Apple M1.)

This has a performance regression (17%) for the single case of Apple M1,
unstriped (P = 0), 8-bit (uint8_t), random input. But for those
conditions P = 1 is almost as fast (within 3%) and unaffected, and P = 0
has very bad performance for uniform data anyway. And this removes the
Apple-specific #pragma unroll, too.

(The performance regression appears to be because in that specific case
Clang unrolls the main loop by a count of 8. Clang is unable to perform
unrolling after the changes here, even with #pragma unroll. We might be
able to recover the speed by semi-manually unrolling by using an
additional loop, or by multiplying the loop count over stripes.)