Manually unroll the loop
This restores (most of) the previous performance regression from adding
the explicit inner loop over stripes (8-bit, P = 0, random input).

On Apple M1, this also benefits 12-bit histograms with P = 0, random
input (and additionally P = 1, random input when filtered). 16-bit was
not affected either way.

The values of NUNROLL here are a first guess and not tuned yet, although
doubling it (to a max of 8) did not do any better on Apple M1.