Avoid branch for filtering
For random input, on Apple M1, this pessimizes the P = 0 case by 11-12%
but improves the P = 2-3 cases by 11-13%, for 9- or 12-bit filtered.

For 14-bit filtered, P = 0 was pessimized by 11-12% and P = 2-3 were
unaffected.

(The pessimization could be shrunk a little by increasing the unrolling
factor from 4 to 8, but I'm not making that change because that factor
has not yet been tuned for Intel.)

The uniform input case was generally very slightly improved.