Roll back some changes for x86-64 only
The reorganization of the loop to be block-wise, as well as the use of
branchless hi-mask filtering, had pessimized x86-64 (Coffee Lake).

Prefer (predictable) branches over branchless on x86-64. Disable extra
unrolling for uint16_t on x86-64.

(For uint8_t, unrolling _is_ beneficial on x86-64, too.)

For x86-64, this is still slower than before introducing the block-wise
loop, in the case of 9/12/14-bit filtered. I suspect that is because the
loop is larger than what we could get when we did not require the block
size to be at least as large as the number of stripes. The fix will
likely be to decouple the samples per iteration ("unrolling") from the
number of stripes to use.

(The separation of bin index computation from the increments does not
appear to have a significant effect, because the compiler optimizes it
back to interleaved code, at least for the single-component case.)