Disable unrolling of unoptimized loop
This keeps code size smaller for the prolog and epilog of the striped
implementation (especially the prolog, where Clang seems unable to see
that it's a short loop, even with a redundant clamp to BLOCKSIZE, and
unrolls it).

As a side effect, the 'unoptimized' implementations will be slower than
they would be if unrolling is left to the compiler, but we probably
don't care too much about that for now.