Remove himask [2]: Remove 'unfiltered' variants
Always use filtering when there are high bits that are not in the range
of the histogram. This is the only safe option, and although it does
have somewhat lower performance than the unfiltered case, I think it is
worth it. It also simplifies everything by eliminating one compile-time
variable.