C API with run-time dispatch
Use a simple C API (ihist.h) for 1/3/4-component 2D images only. Other
patterns may be provided by a future C++ template API, but not for now.

1D functions are probably not necessary in the API unless/until there is
evidence that they can be meaningfully faster.

LoBit is also left out of API unless/until we find out that it's not
YAGNI.

Tuning used for the API functions is temporary (to be supplied for each
platform/arch).

Other changes:
- The implementation function templates in ihist.hpp are (mostly) marked
  noinline, so that the API functions are not bloated.
- Fixed mask bits for random_input tests (was 8, now 1) -- while the
  test was valid either way, 50-50 true-false is what was intended.