Limit threads to number of physical cores
On 6-core Coffee Lake with Hyper-Threading enabled, multithreaded
benchmarks (defaulting to use a concurrency of 12) showed that
efficiency was about half of what it is when the thread count was
limited to 6. This is not surprising given that Hyper-Threading shares
the CPU backend and our histogramming is backend-bound.

Using all logical cores simply wastes CPU time without any speedup, so
limit our thread count to the number of physical cores.

To determine the number of physical cores, 3 methods were considered:

- The hwloc library can provide this information across platforms.
  However, it is cumbersome to build and depend on (Autotools, with
  a CMake build on Windows, but the CMake build produces neither a
  pkg-config or CMake discovery mechanism).

- oneTBB can provide this information via tbb::info::default_concurrency
  (with max_threads_per_core constraint). However, that function just
  returns the number of logical cores unless TBB is built with TBBBind
  support, which in turn depends on hwloc. And TBB requires a specific
  version of hwloc, namely 2.5 as of this writing, which doesn't even
  have the Windows CMake build (only a Visual Studio solution).

- We can just call the OS APIs directly. This ended up being simpler
  than setting up hwloc as a dependency.