Remove oneTBB static library build from justfile
Use as a static library for tests and benchmarks causes problems (at
least on macOS) because templates instantiated from oneTBB headers under
different compiler flags (visibility?) do not get correctly merged at
link time (see 9a53eeff52af8b51e7761ca4ca09d85153300c05; the workaround
used there is not scalable).

So let's just use oneTBB installed on the system (as a shared library)
for testing and benchmarking. This is if anything more convenient on
macOS (Homebrew) and Linux. On Windows we might add a way to
automatically download a oneTBB binary release (and set up paths for
build and run).

(We might resurrect the building of a static oneTBB for future language
bindings, where we shouldn't have a problem if oneTBB is linked only
once, directly.)