Set pom.xml version from meson.build
For now, set in justfile with -SANPSHOT appended. Releases will need to
set a non-snapshot version.

Default to 0.0.0-SNAPSHOT in pom.xml so there will be no accidents if we
fail to set -Drevision (the worst that could happen is we deploy a
snapshot).

(Assisted by Claude Code; any errors are mine.)